+++ neural_lam/__init__.py
@@ -9,6 +9,7 @@ import neural_lam.utils
 import neural_lam.vis
 
 # Local
+from .batch import ForecastBatch
 from .weather_dataset import WeatherDataset
 
 try: